requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
drf-spectacular>=0.22.1,<0.23
uwsgi>=2.0.20,<2.1
psycopg2>=2.9.3,<2.10
django-cors-headers
polars
transformers
pandas
pyarrow
openai
accelerate
torch
celery
redis